+++ zipmanager/__init__.py
@@ -1,1 +1,2 @@
 from .main import ZipFolder
+from .File import create_base